// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      user_missions: {
        Row: {
          id: string;
          user_id: string;
          mission_id: string;
          progress: number;
          status: 'locked' | 'in_progress' | 'completed' | 'claimed';
          completed_at?: string;
          expires_at?: string;
          created_at: string;
        }
        Insert: {
          id: string;
          user_id: string;
          mission_id: string;
          progress: number;
          status: 'locked' | 'in_progress' | 'completed' | 'claimed';
          completed_at?: string;
          expires_at?: string;
          created_at: string;
        }
        Update: {
          id?: string;
          user_id?: string;
          mission_id?: string;
          progress?: number;
          status?: 'locked' | 'in_progress' | 'completed' | 'claimed';
          completed_at?: string;
          expires_at?: string;
          created_at?: string;
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}